requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
tqdm==4.66.1
python-dotenv==1.0.0
validators==0.22.0
